requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
pandas-datareader==0.10.0
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
matplotlib==3.8.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
